Workaround Terraform AWS creds timing out
stscreds.DefaultDuration is 15 mins. This has caused some problems for
longer running builds. Ideally the creds should refresh, but the config
container provides them to the terraform container via environment vars
meaning that they can't be updated. It would probably be necessary to
make terraform itself assume the role, which is possible.